feat: move Velvet class to Velvet folder
