+++ src/index.ts
@@ -1,3 +1,3 @@
-import Velvet from "./Velvet.js";
+import Velvet from "./Velvet/Velvet.js";
 
 export default Velvet;
